Add super class of people_in_school, add parent class of staff to the two teachers, add greetable module because all high five. Add constants to staff class for rating threshhold and teach_stuff responses
