+++ student.rb
@@ -1,15 +1,9 @@
-class Student
-  attr_reader :age, :phase
-  attr_accessor :name
+require_relative 'people_in_school'
+
+class Student < PeopleInSchool
 
   def initialize(options = {})
     @phase = 1
-    @age = options.fetch(:age, 0)
-    @name = options.fetch(:name, "")
-  end
-
-  def offer_high_five
-    "High five!"
   end
 
   def set_phase(num)
